ragcore-doc-analyzer/requirements.txt
fastapi
uvicorn
requests
pydantic
python-dotenv
pyyaml
PyPDF2
unstructured[all-docs]
numpy<2.0